requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pytest==8.2.1
httpx==0.27.0
anyio==4.3.0
pyjwt==2.8.0
flake8==7.2.0
pylint==3.3.6
radon==6.0.1
